+++ src/main/java/com/example/database_user/controllers/api/DiscipuladoMenoresAPI.java
@@ -1,9 +1,9 @@
 package com.example.database_user.controllers.api;
 
-import com.example.database_user.model.dto.DiscipuladoMenoresDTO;
 import com.example.database_user.model.dto.Persona.PersonaDTO;
 import com.example.database_user.model.dto.Persona.PersonaNinos;
 import com.example.database_user.model.dto.Persona.SimplePersona;
+import com.example.database_user.model.dto.discipulado.DiscipuladoMenoresDTO;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
