+++ src/main/java/com/example/database_user/controllers/impl/DiscipuladoMenoresController.java
@@ -2,10 +2,10 @@ package com.example.database_user.controllers.impl;
 
 
 import com.example.database_user.controllers.api.DiscipuladoMenoresAPI;
-import com.example.database_user.model.dto.DiscipuladoMenoresDTO;
 import com.example.database_user.model.dto.Persona.PersonaDTO;
 import com.example.database_user.model.dto.Persona.PersonaNinos;
 import com.example.database_user.model.dto.Persona.SimplePersona;
+import com.example.database_user.model.dto.discipulado.DiscipuladoMenoresDTO;
 import com.example.database_user.services.DiscipuladoMenoresService;
 import jakarta.validation.Valid;
 import java.util.List;
