+++ src/main/java/com/example/database_user/controllers/impl/EmmausController.java
@@ -1,7 +1,7 @@
 package com.example.database_user.controllers.impl;
 
 import com.example.database_user.controllers.api.EmmausAPI;
-import com.example.database_user.model.dto.EmausDTO;
+import com.example.database_user.model.dto.discipulado.EmausDTO;
 import com.example.database_user.services.EmmausService;
 import java.util.List;
 import lombok.AllArgsConstructor;
@@ -12,10 +12,11 @@ import org.springframework.web.bind.annotation.RestController;
 @AllArgsConstructor
 public class EmmausController implements EmmausAPI {
 
-    private final EmmausService emmausService;
-    public ResponseEntity<List<EmausDTO>> getAllEmmaus() {
-        //Do not query the people inside
-        return null;
-    }
+  private final EmmausService emmausService;
+
+  public ResponseEntity<List<EmausDTO>> getAllEmmaus() {
+    //Do not query the people inside
+    return null;
+  }
 
 }
