+++ src/main/java/com/example/database_user/model/dto/Persona/PersonaDTO.java
@@ -1,12 +1,11 @@
 package com.example.database_user.model.dto.Persona;
 
-import com.example.database_user.model.dto.Custom;
+import com.example.database_user.model.dto.Custom.CustomTagDTO;
 import com.example.database_user.model.dto.FileStorage;
 import com.example.database_user.model.dto.PersonGroups;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -18,6 +17,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class PersonaDTO {
+
   private String id;
   private String nombre;
   private String apellido;
@@ -26,7 +26,7 @@ public class PersonaDTO {
   private LocalDate birthday;
   private LocalDate saint;
   private String dni;
-  private List<Custom> extras;
+  private List<CustomTagDTO> extras;
   private List<PersonaSacramentsDTO> sacraments;
   private List<FileStorage> fileStorage;
   private LocalDateTime createdOn;
