+++ src/main/java/com/example/database_user/model/mapper/DiscipuladoMenoresMapper.java
@@ -1,13 +1,14 @@
 package com.example.database_user.model.mapper;
 
-import com.example.database_user.model.dto.DiscipuladoMenoresDTO;
+import com.example.database_user.model.dto.discipulado.DiscipuladoMenoresDTO;
 import com.example.database_user.repositories.entity.DiscipuladoMenoresEntity;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring")
 public interface DiscipuladoMenoresMapper {
-    DiscipuladoMenoresDTO toDTO(DiscipuladoMenoresEntity persona);
 
-    DiscipuladoMenoresEntity toEntity(DiscipuladoMenoresDTO persona);
+  DiscipuladoMenoresDTO toDTO(DiscipuladoMenoresEntity persona);
+
+  DiscipuladoMenoresEntity toEntity(DiscipuladoMenoresDTO persona);
 
 }
