+++ src/main/java/com/example/database_user/services/DiscipuladoMenoresService.java
@@ -1,47 +1,48 @@
 package com.example.database_user.services;
 
-import com.example.database_user.model.dto.DiscipuladoMenoresDTO;
 import com.example.database_user.model.dto.Persona.PersonaDTO;
 import com.example.database_user.model.dto.Persona.SimplePersona;
+import com.example.database_user.model.dto.discipulado.DiscipuladoMenoresDTO;
+import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 @Service
 public interface DiscipuladoMenoresService {
-    ResponseEntity<DiscipuladoMenoresDTO> getDiscipuladoMenoresById(String id);
 
-    ResponseEntity<List<DiscipuladoMenoresDTO>> fetchAllDiscipuladoMenores();
+  ResponseEntity<DiscipuladoMenoresDTO> getDiscipuladoMenoresById(String id);
 
-    ResponseEntity<List<SimplePersona>> getDiscipuladoMenoresInscritosById(String id);
+  ResponseEntity<List<DiscipuladoMenoresDTO>> fetchAllDiscipuladoMenores();
 
-    ResponseEntity<List<PersonaDTO>> getDiscipuladoMenoresMonisById(String id);
+  ResponseEntity<List<SimplePersona>> getDiscipuladoMenoresInscritosById(String id);
 
-    ResponseEntity<String> insertNewDiscipuladoMenores(
-            DiscipuladoMenoresDTO discipuladoMenoresDTO);
+  ResponseEntity<List<PersonaDTO>> getDiscipuladoMenoresMonisById(String id);
 
-    ResponseEntity<String> deleteDiscipuladoMenoresById(String id);
+  ResponseEntity<String> insertNewDiscipuladoMenores(
+      DiscipuladoMenoresDTO discipuladoMenoresDTO);
 
-    ResponseEntity<String> updateDiscipuladoMenores(DiscipuladoMenoresDTO discipuladoMenoresDTO);
+  ResponseEntity<String> deleteDiscipuladoMenoresById(String id);
 
-    ResponseEntity<String> addNewUserNewInsciption(
-            PersonaDTO personaDTO,
-            String idDiscipuladoMenores);
+  ResponseEntity<String> updateDiscipuladoMenores(DiscipuladoMenoresDTO discipuladoMenoresDTO);
 
-    ResponseEntity<String> addExistingUserExistingInscription(
-            String idDiscipuladoMenores,
-            String idPerson);
+  ResponseEntity<String> addNewUserNewInsciption(
+      PersonaDTO personaDTO,
+      String idDiscipuladoMenores);
 
-    ResponseEntity<String> addExistingUserNewInscription(
-            PersonaDTO personaDTO,
-            String idDiscipuladoMenores);
+  ResponseEntity<String> addExistingUserExistingInscription(
+      String idDiscipuladoMenores,
+      String idPerson);
 
-    ResponseEntity<String> editExistingUserInscription(PersonaDTO personaDTO);
+  ResponseEntity<String> addExistingUserNewInscription(
+      PersonaDTO personaDTO,
+      String idDiscipuladoMenores);
 
-    ResponseEntity<String> deleteExistingInscriptionFromADiscipuladoMenores(
-            String idDiscipuladoMenores, String idPersona);
+  ResponseEntity<String> editExistingUserInscription(PersonaDTO personaDTO);
 
-    Integer countInscritos(String idDiscipuladoMenores);
+  ResponseEntity<String> deleteExistingInscriptionFromADiscipuladoMenores(
+      String idDiscipuladoMenores, String idPersona);
+
+  Integer countInscritos(String idDiscipuladoMenores);
 }
